+++ packages/core/src/index.ts
@@ -2,6 +2,7 @@ export type { ClientClass } from './sdk';
 export type { Carrier, Layer } from './hub';
 export type { OfflineStore, OfflineTransportOptions } from './transports/offline';
 
+export * from './tracing';
 export {
   addBreadcrumb,
   captureException,
