+++ packages/core/src/tracing/errors.ts
