+++ packages/core/src/tracing/index.ts
@@ -1,0 +1,7 @@
+export * from './errors';
+export * from './hubextensions';
+export * from './idletransaction';
+export * from './span';
+export * from './transaction';
+export * from './utils';
+export * from './spanstatus';
