+++ packages/core/src/tracing/span.ts
