+++ packages/tracing/src/browser/browsertracing.ts
@@ -1,12 +1,15 @@
 /* eslint-disable max-lines */
-import type { Hub } from '@sentry/core';
+import type { Hub, IdleTransaction } from '@sentry/core';
+import {
+  DEFAULT_FINAL_TIMEOUT,
+  DEFAULT_HEARTBEAT_INTERVAL,
+  DEFAULT_IDLE_TIMEOUT,
+  extractTraceparentData,
+  startIdleTransaction,
+} from '@sentry/core';
 import type { EventProcessor, Integration, Transaction, TransactionContext, TransactionSource } from '@sentry/types';
 import { baggageHeaderToDynamicSamplingContext, getDomElement, logger } from '@sentry/utils';
 
-import { startIdleTransaction } from '../hubextensions';
-import type { IdleTransaction } from '../idletransaction';
-import { DEFAULT_FINAL_TIMEOUT, DEFAULT_HEARTBEAT_INTERVAL, DEFAULT_IDLE_TIMEOUT } from '../idletransaction';
-import { extractTraceparentData } from '../utils';
 import { registerBackgroundTabDetection } from './backgroundtab';
 import { addPerformanceEntries, startTrackingLongTasks, startTrackingWebVitals } from './metrics';
 import type { RequestInstrumentationOptions } from './request';
