+++ packages/tracing/src/index.bundle.ts
@@ -53,13 +53,13 @@ export {
 export { SDK_VERSION } from '@sentry/browser';
 
 import { Integrations as BrowserIntegrations } from '@sentry/browser';
+import { addExtensionMethods } from '@sentry/core/src/tracing/hubextensions';
 import type { Integration } from '@sentry/types';
 import { GLOBAL_OBJ } from '@sentry/utils';
 
 import { BrowserTracing } from './browser';
-import { addExtensionMethods } from './hubextensions';
 
-export { Span } from './span';
+export { Span } from '@sentry/core/src/tracing/span';
 
 let windowIntegrations = {};
