+++ packages/tracing/tsconfig.json
@@ -1,8 +1,16 @@
 {
   "extends": "../../tsconfig.json",
-
-  "include": ["src/**/*"],
-
+  "include": [
+    "src/**/*",
+    "../core/src/tracing/errors.ts",
+    "../core/src/tracing/errors.ts",
+    "../core/src/tracing/hubextensions.ts",
+    "../core/src/tracing/idletransaction.ts",
+    "../core/src/tracing/span.ts",
+    "../core/src/tracing/spanstatus.ts",
+    "../core/src/tracing/transaction.ts",
+    "../core/src/tracing/utils.ts"
+  ],
   "compilerOptions": {
     // package-specific options
   }
